Change v2 and ignore

+++ internal/mainservice/models/client.go
@@ -4,7 +4,7 @@ type ClientRegister struct {
 	Name       string `json:"name"`
 	ClientUuid string `json:"clientuuid"`
 	Address    string `json:"address"`
-	Phone      int    `json:"phone"`
+	Phone      int64  `json:"phone"`
 	DayWork    string `json:"daywork"`
 	UserEmail  string
 }
@@ -12,6 +12,6 @@ type ClientResponce struct {
 	Name       string `json:"name"`
 	ClientUuid string `json:"uuid"`
 	Address    string `json:"address"`
-	Phone      int    `json:"phone"`
+	Phone      int64  `json:"phone"`
 	DayWork    string `json:"daywork"`
 }
